+++ examples/cookbook/app/custom-render/__tests__/index.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { screen } from '@testing-library/react-native';
 import { renderWithProviders } from './test-utils';
-import WelcomeScreen from "../../../app/custom-render";
+import WelcomeScreen from '../WelcomeScreen';
 
 test('renders WelcomeScreen in light theme', () => {
   renderWithProviders(<WelcomeScreen />, { theme: 'light' });
